--- a/.github/actions/next-stats-action/src/prepare/repo-setup.js
+++ b/.github/actions/next-stats-action/src/prepare/repo-setup.js
@@ -4,6 +4,7 @@ const { existsSync } = require('fs')
 const exec = require('../util/exec')
 const logger = require('../util/logger')
 const execa = require('execa')
+const { Span } = require('next/dist/trace')
 
 module.exports = (actionInfo) => {
   return {
@@ -59,7 +60,11 @@ module.exports = (actionInfo) => {
      * @param {{ repoDir: string, nextSwcVersion: null | string }} options Required options
      * @returns {Promise<Map<string, string>>} List packages key is the package name, value is the path to the packed tar file.'
      */
-    async linkPackages({ repoDir, nextSwcVersion }) {
+    async linkPackages({ repoDir, nextSwcVersion, parentSpan }) {
+      if (!parentSpan) {
+        // Not all callers provide a parent span
+        parentSpan = new Span({ name: 'dummy' })
+      }
       /** @type {Map<string, string>} */
       const pkgPaths = new Map()
       /** @type {Map<string, { packageJsonPath: string, packagePath: string, packageJson: any, packedPackageTarPath: string }>} */
@@ -68,9 +73,11 @@ module.exports = (actionInfo) => {
       let packageFolders
 
       try {
-        packageFolders = await fs.promises.readdir(
-          path.join(repoDir, 'packages')
-        )
+        packageFolders = await parentSpan
+          .traceChild('read-packages-folder')
+          .traceAsyncFn(() =>
+            fs.promises.readdir(path.join(repoDir, 'packages'))
+          )
       } catch (err) {
         if (err.code === 'ENOENT') {
           require('console').log('no packages to link')
@@ -79,162 +86,190 @@ module.exports = (actionInfo) => {
         throw err
       }
 
-      for (const packageFolder of packageFolders) {
-        const packagePath = path.join(repoDir, 'packages', packageFolder)
-        const packedPackageTarPath = path.join(
-          packagePath,
-          `${packageFolder}-packed.tgz`
-        )
-        const packageJsonPath = path.join(packagePath, 'package.json')
-
-        if (!existsSync(packageJsonPath)) {
-          require('console').log(`Skipping ${packageFolder}, no package.json`)
-          continue
-        }
-
-        const packageJson = JSON.parse(fs.readFileSync(packageJsonPath))
-        const { name: packageName } = packageJson
-
-        pkgDatas.set(packageName, {
-          packageJsonPath,
-          packagePath,
-          packageJson,
-          packedPackageTarPath,
-        })
-        pkgPaths.set(packageName, packedPackageTarPath)
-      }
-
-      for (const [
-        packageName,
-        { packageJsonPath, packagePath, packageJson },
-      ] of pkgDatas.entries()) {
-        // This loops through all items to get the packagedPkgPath of each item and add it to pkgData.dependencies
-        for (const [
-          packageName,
-          { packedPackageTarPath },
-        ] of pkgDatas.entries()) {
-          if (
-            !packageJson.dependencies ||
-            !packageJson.dependencies[packageName]
+      parentSpan.traceChild('get-pkgdatas').traceFn(() => {
+        for (const packageFolder of packageFolders) {
+          const packagePath = path.join(repoDir, 'packages', packageFolder)
+          const packedPackageTarPath = path.join(
+            packagePath,
+            `${packageFolder}-packed.tgz`
           )
+          const packageJsonPath = path.join(packagePath, 'package.json')
+
+          if (!existsSync(packageJsonPath)) {
+            require('console').log(`Skipping ${packageFolder}, no package.json`)
             continue
-          // Edit the pkgData of the current item to point to the packed tgz
-          packageJson.dependencies[packageName] = packedPackageTarPath
-        }
-
-        // make sure native binaries are included in local linking
-        if (packageName === '@next/swc') {
-          packageJson.files ||= []
-
-          packageJson.files.push('native')
-
-          try {
-            const swcBinariesDirContents = (
-              await fs.promises.readdir(path.join(packagePath, 'native'))
-            ).filter((file) => file !== '.gitignore' && file !== 'index.d.ts')
-
-            require('console').log(
-              'using swc binaries: ',
-              swcBinariesDirContents.join(', ')
-            )
-          } catch (err) {
-            if (err.code === 'ENOENT') {
-              require('console').log('swc binaries dir is missing!')
-            }
-            throw err
           }
-        } else if (packageName === 'next') {
-          const nextSwcPkg = pkgDatas.get('@next/swc')
 
-          console.log('using swc dep', {
-            nextSwcVersion,
-            nextSwcPkg,
+          const packageJson = JSON.parse(fs.readFileSync(packageJsonPath))
+          const { name: packageName } = packageJson
+
+          pkgDatas.set(packageName, {
+            packageJsonPath,
+            packagePath,
+            packageJson,
+            packedPackageTarPath,
           })
-          if (nextSwcVersion) {
-            Object.assign(packageJson.dependencies, {
-              '@next/swc-linux-x64-gnu': nextSwcVersion,
-            })
-          } else {
-            if (nextSwcPkg) {
-              packageJson.dependencies['@next/swc'] =
-                nextSwcPkg.packedPackageTarPath
-            }
-          }
+          pkgPaths.set(packageName, packedPackageTarPath)
         }
+      })
 
-        await fs.promises.writeFile(
-          packageJsonPath,
-          JSON.stringify(packageJson, null, 2),
-          'utf8'
-        )
-      }
-
-      // wait to pack packages until after dependency paths have been updated
-      // to the correct versions
-      await Promise.all(
-        Array.from(pkgDatas.entries()).map(
-          async ([
+      await parentSpan
+        .traceChild('write-packagejson')
+        .traceAsyncFn(async () => {
+          for (const [
             packageName,
-            { packagePath: pkgPath, packedPackageTarPath: packedPkgPath },
-          ]) => {
-            /** @type {null | () => Promise<void>} */
-            let cleanup = null
+            { packageJsonPath, packagePath, packageJson },
+          ] of pkgDatas.entries()) {
+            // This loops through all items to get the packagedPkgPath of each item and add it to pkgData.dependencies
+            for (const [
+              packageName,
+              { packedPackageTarPath },
+            ] of pkgDatas.entries()) {
+              if (
+                !packageJson.dependencies ||
+                !packageJson.dependencies[packageName]
+              )
+                continue
+              // Edit the pkgData of the current item to point to the packed tgz
+              packageJson.dependencies[packageName] = packedPackageTarPath
+            }
 
+            // make sure native binaries are included in local linking
             if (packageName === '@next/swc') {
-              // next-swc uses a gitignore to prevent the committing of native builds but it doesn't
-              // use files in package.json because it publishes to individual packages based on architecture.
-              // When we used yarn to pack these packages the gitignore was ignored so the native builds were packed
-              // however npm does respect gitignore when packing so we need to remove it in this specific case
-              // to ensure the native builds are packed for use in gh actions and related scripts
+              packageJson.files ||= []
 
-              const nativeGitignorePath = path.join(
-                pkgPath,
-                'native/.gitignore'
-              )
-              const renamedGitignorePath = path.join(
-                pkgPath,
-                'disabled-native-gitignore'
-              )
+              packageJson.files.push('native')
 
-              await fs.promises.rename(
-                nativeGitignorePath,
-                renamedGitignorePath
-              )
-              cleanup = async () => {
-                await fs.promises.rename(
-                  renamedGitignorePath,
-                  nativeGitignorePath
+              try {
+                const swcBinariesDirContents = (
+                  await fs.promises.readdir(path.join(packagePath, 'native'))
+                ).filter(
+                  (file) => file !== '.gitignore' && file !== 'index.d.ts'
                 )
+
+                require('console').log(
+                  'using swc binaries: ',
+                  swcBinariesDirContents.join(', ')
+                )
+              } catch (err) {
+                if (err.code === 'ENOENT') {
+                  require('console').log('swc binaries dir is missing!')
+                }
+                throw err
+              }
+            } else if (packageName === 'next') {
+              const nextSwcPkg = pkgDatas.get('@next/swc')
+
+              console.log('using swc dep', {
+                nextSwcVersion,
+                nextSwcPkg,
+              })
+              if (nextSwcVersion) {
+                Object.assign(packageJson.dependencies, {
+                  '@next/swc-linux-x64-gnu': nextSwcVersion,
+                })
+              } else {
+                if (nextSwcPkg) {
+                  // packageJson.dependencies['@next/swc'] =
+                  //   nextSwcPkg.packedPackageTarPath
+                }
               }
             }
 
-            const options = {
-              cwd: pkgPath,
-              env: {
-                ...process.env,
-                COREPACK_ENABLE_STRICT: '0',
-              },
-            }
-            let execResult
-            try {
-              execResult = await execa('pnpm', ['pack'], options)
-            } catch {
-              execResult = await execa('pnpm', ['pack'], options)
-            }
-            const { stdout } = execResult
-
-            const packedFileName = stdout.trim()
-
-            await Promise.all([
-              fs.promises.rename(
-                path.join(pkgPath, packedFileName),
-                packedPkgPath
-              ),
-              cleanup?.(),
-            ])
+            await fs.promises.writeFile(
+              packageJsonPath,
+              JSON.stringify(packageJson, null, 2),
+              'utf8'
+            )
           }
-        )
-      )
+        })
+
+      await parentSpan
+        .traceChild('pnpm-packing')
+        .traceAsyncFn(async (packingSpan) => {
+          // wait to pack packages until after dependency paths have been updated
+          // to the correct versions
+          await Promise.all(
+            Array.from(pkgDatas.entries()).map(
+              async ([
+                packageName,
+                { packagePath: pkgPath, packedPackageTarPath: packedPkgPath },
+              ]) => {
+                return packingSpan
+                  .traceChild('handle-package', { packageName })
+                  .traceAsyncFn(async (handlePackageSpan) => {
+                    /** @type {null | () => Promise<void>} */
+                    let cleanup = null
+
+                    if (packageName === '@next/swc') {
+                      // next-swc uses a gitignore to prevent the committing of native builds but it doesn't
+                      // use files in package.json because it publishes to individual packages based on architecture.
+                      // When we used yarn to pack these packages the gitignore was ignored so the native builds were packed
+                      // however npm does respect gitignore when packing so we need to remove it in this specific case
+                      // to ensure the native builds are packed for use in gh actions and related scripts
+
+                      const nativeGitignorePath = path.join(
+                        pkgPath,
+                        'native/.gitignore'
+                      )
+                      const renamedGitignorePath = path.join(
+                        pkgPath,
+                        'disabled-native-gitignore'
+                      )
+
+                      await handlePackageSpan
+                        .traceChild('rename-gitignore')
+                        .traceAsyncFn(() =>
+                          fs.promises.rename(
+                            nativeGitignorePath,
+                            renamedGitignorePath
+                          )
+                        )
+                      cleanup = async () => {
+                        await fs.promises.rename(
+                          renamedGitignorePath,
+                          nativeGitignorePath
+                        )
+                      }
+                    }
+
+                    const options = {
+                      cwd: pkgPath,
+                      env: {
+                        ...process.env,
+                        COREPACK_ENABLE_STRICT: '0',
+                      },
+                    }
+                    let execResult
+                    try {
+                      execResult = await handlePackageSpan
+                        .traceChild('pnpm-pack-try-1')
+                        .traceAsyncFn(() => execa('pnpm', ['pack'], options))
+                    } catch {
+                      execResult = await handlePackageSpan
+                        .traceChild('pnpm-pack-try-2')
+                        .traceAsyncFn(() => execa('pnpm', ['pack'], options))
+                    }
+                    const { stdout } = execResult
+
+                    const packedFileName = stdout.trim()
+
+                    await handlePackageSpan
+                      .traceChild('rename-packed-tar-and-cleanup')
+                      .traceAsyncFn(() =>
+                        Promise.all([
+                          fs.promises.rename(
+                            path.join(pkgPath, packedFileName),
+                            packedPkgPath
+                          ),
+                          cleanup?.(),
+                        ])
+                      )
+                  })
+              }
+            )
+          )
+        })
 
       return pkgPaths
     },