@ijjk 2024-04-23
## Failing test suites <!--23052555760-->
Commit: 93a4d3851aa95cb25465b5fe7529aaf4e69d7b79

`pnpm test-start test/e2e/app-dir/app-compilation/index.test.ts`
- app dir > Loading > should render loading.js in initial html for slow page

<details>
<summary>Expand output</summary>

  ● app dir › Loading › should render loading.js in initial html for slow page

    next build failed with code/signal 1

      89 |           if (code || signal)
      90 |             reject(
    > 91 |               new Error(`next build failed with code/signal ${code || signal}`)
         |               ^
      92 |             )
      93 |           else resolve()
      94 |         })

      at ChildProcess.<anonymous> (lib/next-modes/next-start.ts:91:15)

</details>


Read more about building and testing Next.js in [contributing.md](https://github.com/vercel/next.js/blob/canary/contributing.md).