@ijjk 2024-04-23
## Failing test suites <!--23052555760-->
Commit: 93a4d3851aa95cb25465b5fe7529aaf4e69d7b79

`__NEXT_EXPERIMENTAL_PPR=true pnpm test-start test/e2e/app-dir/actions-allowed-origins/app-action-disallowed-origins.test.ts` (PPR)
- app-dir action disallowed origins > should error if x-forwarded-host does not match the origin

<details>
<summary>Expand output</summary>

  ● app-dir action disallowed origins › should error if x-forwarded-host does not match the origin

    next build failed with code/signal 1

      89 |           if (code || signal)
      90 |             reject(
    > 91 |               new Error(`next build failed with code/signal ${code || signal}`)
         |               ^
      92 |             )
      93 |           else resolve()
      94 |         })

      at ChildProcess.<anonymous> (lib/next-modes/next-start.ts:91:15)

</details>


Read more about building and testing Next.js in [contributing.md](https://github.com/vercel/next.js/blob/canary/contributing.md).

`__NEXT_EXPERIMENTAL_PPR=true pnpm test-start test/e2e/app-dir/app-basepath-custom-server/index.test.ts` (PPR)
- custom-app-server-action-redirect > redirects with basepath properly when server action handler uses `redirect`
- custom-app-server-action-redirect > redirects with proper cookies set from both redirect response and post respose

<details>
<summary>Expand output</summary>

  ● custom-app-server-action-redirect › redirects with basepath properly when server action handler uses `redirect`

    next build failed with code/signal 1

      89 |           if (code || signal)
      90 |             reject(
    > 91 |               new Error(`next build failed with code/signal ${code || signal}`)
         |               ^
      92 |             )
      93 |           else resolve()
      94 |         })

      at ChildProcess.<anonymous> (lib/next-modes/next-start.ts:91:15)

  ● custom-app-server-action-redirect › redirects with proper cookies set from both redirect response and post respose

    next build failed with code/signal 1

      89 |           if (code || signal)
      90 |             reject(
    > 91 |               new Error(`next build failed with code/signal ${code || signal}`)
         |               ^
      92 |             )
      93 |           else resolve()
      94 |         })

      at ChildProcess.<anonymous> (lib/next-modes/next-start.ts:91:15)

</details>


Read more about building and testing Next.js in [contributing.md](https://github.com/vercel/next.js/blob/canary/contributing.md).